Fixed infinite loop involving the scheduling
Changed the simulation loop condition to be based on the world clock instead of events in the priority queue since there will always be a new car arriving event before a car departs the garage. (Since a new car arrives between 1 and 8 minutes and a car departs between 1 and 50 minutes).